Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> I'm curious, why are you still using gclient-new-workdir at all? I wouldn't have thought it was still useful.

I'm working on a project which is built on top of chromium. When we start working on updating to a new chrome dev release, the beta release is not yet finished, so it's very convenient to keep different working trees at the same time.

Patch-set: 5
Reviewer: Gerrit User 1244248 <1244248@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>