Update patch set 4
Patch Set 4:

The approach seems sane, but when I patched this into my depot_tools and then fetched chrome, it still didn't print progress. I also tried wrapping the WriteToStdout wrapper in an AutoFlush wrapper, but that didn't help. As far as I can tell, the output from the clone command is still only going to the in-memory cStringIO buffer, and only gets printed to stdout when I <ctrl-c> the process and it prints all output so far while recovering from the exception.

It might be something to do with the gclient_utils.CheckCallAndFilter, and the default gclient_utils.GitFilter. It doesn't seem like those two should be breaking it (they're *supposed* to print one progress line per second, by default, and as far as I can tell from the output when I <ctrl-c> the process, they seem to be doing the right thing) but they're definitely a complication.

Patch-set: 4
Reviewer: Gerrit User 1002100 <1002100@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>