Update patch set 6
Patch Set 6: Code-Review+1

Great! Tested this with chromium and with infra and it does basically what we want. It appears that it is willing to print to stdout more than once per second, which I would consider fixing, but this is a massive improvement. Remove those comments and ship it!

Patch-set: 6
Reviewer: Gerrit User 1002100 <1002100@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1