git_test_utils: Fixed custom git metadata in git commit schemas.
Previously, this feature was somewhat implemented but didn't work at
all. I have no idea what this code was trying to do, but it didn't work
(and isn't used in any tests).

Fixed the implementation so now if you specify custom keys on a commit
schema (like 'AUTHOR_NAME'), that will actually be set as the correct
git metadata.

Bug: 808941
Change-Id: Id7c4fa5822741925beba591ea587bd8ebbf2e478