bot_update: use cherry-pick to allow patching onto an older revision.
This assumes that the ref fetched from Gerrit always has the
patchset as 1 commit on top of the base commit. I think this is
correct because in Gerrit's UI the "Download" dialog endorses the
usage of `git cherry-pick FETCH_HEAD`.

This is an alternative (heavily modified reland of)
I68e4d805811530b585550bc75099354fef4e9c15

Bug: chromium:809892
Change-Id: I802c860104a71e7ceb623962f858416e8a9a7e08