ast + tokenize Demo
This also attempts to save the necessary information together with the
dictionary parsed by gclient_eval.

It does so by parsing while evaluating the schema.

I used an ugly hack to create an object that behaves like a string, but
has a node attribute with the information necessary to edit without
losing formatting and comments.

We could also use tuples or objects, but that would imply changing the
client code.

Bug: None
Change-Id: I6a1c77086ae218f6d7a464c45d18b297d4dec065