ast + tokenize Demo 2
This also attempts to save the necessary information together with the
dictionary parsed by gclient_eval.

It does so by using a hack to add a '.node' attribute with the ast
node for the parsed python dict, while maintaining existing behaviour.

It also adds EditRevision and EditVar functions to support editing
git revisions and variables.

Bug: None
Change-Id: I6a1c77086ae218f6d7a464c45d18b297d4dec065