Refactor sync_submodules recipe module into depot_tools
This brings the core of infra.git's sync_submodules recipe modules
(the part which actually runs deps2submodules.py) into depot_tools,
while leaving the fluff (getting a particular set of checkouts in
particular places on disc) in the old module. This enables us to
refactor the deps2submodules.py resource into something that
directly relies on gclient's DEPS-parsing library, rather than
poorly reimplementing it.

This CL will be followed by one to sync_submodules in infra.git
to make it rely on this, and then a refactoring of this module
to change the way deps2submodules.py parses DEPS files.

Bug: 818798
Change-Id: I468d032a701c366c6991610227d44e2dd1ee8f73