ast + tokenize Demo 3
We can also modify the schema to accept things-that-behave-like-dicts
too. We only need a custom dict type that stores the tokens and ast
information.

What I like about this is that we can use the schema to validate the
format, and use the resulting dict to navigate to the dep we want to
modify.

Bug: None
Change-Id: I6c0712cc079dbbbaee6541b7eda71f4b4813b77b