Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > This naively seems like a terrible idea, and I'm a bit confused on the motivation. Can you provide a more concrete example of what goes wrong?
> 
> Sure can: https://uberchromegw.corp.google.com/i/official.infra.cron/builders/chrome-branch/builds/467
> 
> Specifically, the error is
> 
>   ________ running 'cipd ensure -log-level error -root /mnt/data/b/build/slave/chrome-branch/build.dead/flatten_checkout/buildspecs -ensure-file /tmp/tmpJE876X.ensure' in '.'
>   Error: failed to parse desired state (line 1): bad subdir: invalid ".": "../src/chrome/android/profiles".
> 
> see also https://bugs.chromium.org/p/chromium/issues/detail?id=755920#c25
> 
> > 
> > Why doesn't gclient's normal handling of paths (all paths in all DEPS files are considered to be relative to the .gclient file in the gclient root, unless use_relative_paths is True) handle this correctly?
> 
> the DEPS file is buildspecs/DEPS, so the cipd packages assume that their cipd root is buildspecs. the dep path is src/foo/bar, though, which -- from the cipd root -- is ../src/foo/bar.

Ok so for a concrete, minimal example:

Suppose we have this file structure:
/.gclient
/src/*
/buildspec/DEPS

Where the /buildspec/DEPS file has the following content:
deps {
  "src/foo": {
    "url": "{chromium_url}/foo.git@{foo_revision}",
  },
  "src/bar": {
    "packages": [
      {
        "package": "chromium/bar",
        "version": "version:1",
      },
    ],
    "dep_type": "cipd",
  }
}

Then the "src/foo" git repo is going to be checked out to /src/foo, even though it is specified in /buildspec/DEPS; it won't be checked out to /buildspec/src/foo, since the DEPS file doesn't specify that relative paths should be used. The cipd package should do the same.

Is the problem that it *tries* to do the same, but in doing so, puts a ".." in the path passed to cipd_ensure, and that causes cipd_ensure to blow up?

Seems like the only real fix here is for cipd and gclient to have the same root?

Patch-set: 1
Reviewer: Gerrit User 1002100 <1002100@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>