Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> If I understand this CL correctly, this kinda feels like the wrong fix to me. 
> 
> It looks like this changes things from "always use python" to "always use vpython". And, it seems like the fix should be "use vpython if I started w/ vpython, use python if I started w/ python"?

Yes, I could do that too, it was just more complicated to implement, and I wasn't sure that the complexity was worth it (since all the presubmit stuff assumes depot_tools anyhow, so any environment that it operates in is going to have vpython too).

The theory is that 'always use vpython' is equivalent to 'always use SOME pinned environment, even if it is empty' to avoid having things accidentally work (e.g. locally, because of a package installed to the system), but then fail on bots or other folks' machines.

If you prefer, I can rework this to do python-if-python and vpython-if-vpython.

Patch-set: 4
Reviewer: Gerrit User 1001907 <1001907@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0