git cl believes I'm working inside GCE
Looks like the error codes issued by python's http library might vary
between versions/distros. The code tries to reach an URL which is only valid
inside GCE and in the event of getting an error it assumes the user is
a non-googler. However in my case, I am not getting the expected error so it
assumes I'm inside GCE. Adding an extra error code seems to fix it.

Change-Id: I08d3656b25d9046f8b2fc0da2a2b7c7a698c1d92