Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I'm curious, was this working properly before?

The previous behavior was incorrect in that it never respected use_relative_paths, but it was easy to work around by providing the full path to roll-dep.  For example, Skia's DEPS file has:

...
use_relative_paths = True
...
  "third_party/externals/skcms": "...",
...

But we had to run roll-dep like this:

$ roll-dep skia/third_party/externals/skcms --roll-to XYZ

in order to to work around this.  Now that deps_path is used directly to look up the dependency in DEPS, the above fails, since there's no "skia/" prefix in that DEPS file.  So I changed the roller to specify the correct deps_path to match what's in DEPS, but that caused errors because roll-dep was assuming that deps_path was relative to root, rather than DEPS.

Patch-set: 1
Reviewer: Gerrit User 1118354 <1118354@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0