Update patch set 1
Patch Set 1:

From [1] and [2] it looks to me that the intent of the code is not to do this (and I'd prefer not to. I think it should be possible to interpret each DEPS file on its own). 

If we end up changing this, we'll need to change those two places (and probably a few more).

[1] https://cs.chromium.org/chromium/tools/depot_tools/gclient.py?l=2294
[2] https://cs.chromium.org/chromium/tools/depot_tools/gclient.py?l=795

Patch-set: 1
CC: Gerrit User 1167909 <1167909@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>