Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > From [1] and [2] it looks to me that the intent of the code is not to do this (and I'd prefer not to. I think it should be possible to interpret each DEPS file on its own). 
> > 
> > If we end up changing this, we'll need to change those two places (and probably a few more).
> > 
> > [1] https://cs.chromium.org/chromium/tools/depot_tools/gclient.py?l=2294
> > [2] https://cs.chromium.org/chromium/tools/depot_tools/gclient.py?l=795
> 
> I added [1], and I'm pretty sure ignoring parent vars wasn't my intent there. That's actually part of flattening, which explicitly combines all the vars from the whole recursedeps hierarchy.

Ok.

What I meant is:

The part that says "DEPS conditionals shouldn't be using vars that aren't also defined in the DEPS", and the fact that we only replace parent variables that were defined in the current DEPS file makes me think that we shouldn't try to use variables that were not defined in the current DEPS file.

So if we decide that we want to use any variables defined by a parent DEPS file, even if they were not declared in the current DEPS file, we should update the comment to say something like "... in the DEPS file *or a parent DEPS file*", and maybe replace [2] with something like

  if self.parent: self._vars.update(self.parent.get_vars())

Patch-set: 2