Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> what if, instead, EvaluateCondition() just returned the original string if it seemed like it just evaluated to itself?
> 
> Would that break the contract of "EvaluateCondition() always returns a bool"?

I think it already does something like that. If it finds a string that's not in vars, it returns it as long as it could be interpreted as a Python name.
It doesn't work for strings that are not Python names (like "amd64-generic", because it tries to interpret it as 'amd64' - 'generic').

Patch-set: 3
Reviewer: Gerrit User 1167909 <1167909@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>