Update patch set 1
Patch Set 1:

Adding a -f option is reasonable, but using argparse to do it doesn't work, I don't think. That's because post_build_ninja_summary.py is designed to be invoked from autoninja.bat which means it needs to be able to handle (and ignore, mostly) all of the arguments that might be passed to ninja. Maybe argparse can handle this, but I think by default it doesn't, and it might be misleading.

Patch-set: 1
Reviewer: Gerrit User 1128439 <1128439@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>