Set core.longpaths to true for Windows bot checkouts.
I'm doing this to fix a bug with the Windows Codesearch builder where git fails to checkout a file with a long path. I've tested this by running the git config command manually on the Windows builder, and have confirmed that it fixes the issue.

The failure in question is here: https://ci.chromium.org/buildbot/chromium.infra.codesearch/codesearch-gen-chromium-win/2

Bug: 396230
Change-Id: Ibf0a0990e6837ae5022a657fb177d5fed4044bb0