Update patch set 3
Patch Set 3:

FWIW, I think the clank bots are wrong to be passing "--revision src/clank@refs/remotes/origin/master" to bot_update. Any idea where that's coming from? Somewhere in the recipes?

It should probably be passing "refs/heads/master" or maybe just "master". The "remotes" syntax is dependent on the way the local repo is configured, and while there are defaults that can often be assumed, there is no guarantee the mapping will always be done in a particular way. We explicitly don't use "remotes" syntax in DEPS files for just this reason, and I think the revisions passed to bot_update or gclient should be the same (i.e. always use the ref name as it exists on the gerrit server).

Patch-set: 3
Reviewer: Gerrit User 1002222 <1002222@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>