Improve the logic for detecting incremental builds
post_build_ninja_summary.py detects a new incremental build by looking
for the end time stamps to go backwards. This means that repeated builds
that have a single long step will not be reliably detect and will, in
fact, be lost entirely. This hits me sometimes when doing link testing -
delete an output DLL, rebuild, and the old results are displayed again.

Checking for a different duration for an existing record as well as an
earlier end time should handle almost all cases, although an explicit
new-build record would be even better.

This also renames the targets local variable to avoid confusion due
to the targets member variable of the Targets class.

Bug: chromium:787983
Change-Id: I60a371df75d6cb7a55bd46b38156cb109feb8f15