Print access errors using logging.error
https://chromium-review.googlesource.com/c/chromium/tools/depot_tools/+/1058722
changed the behaviour of reporting access errors to no longer
interleave them throughout the output, but instead batch them
together.

This CL had the side effect of changing the way the errors are printed
from using logging.error to print. This means that the errors are now
printing to stdout instead of stderr. This is causing personal scripts
I have for generating reports to fail, since they filtered out stderr.

Change-Id: I664f6166c55f8d20d3c35b63eda695335ab7f8a6