Update patch set 5
Patch Set 5:

> Patch Set 5: Code-Review+1
> 
> None URLs in the .gclient file?

Yes, that or in the equivalent "--spec" flag, like:
  gclient sync --spec 'solutions = [{"name": "newbuildspec", "deps_file": "DEPS", "url": None}]'
to get a checkout based on the buildspec represented by "$PWD/newbuildspec/DEPS". That's just one example, but it's used like that a lot. Also, IIRC, some internal team is doing something similar to parse DEPS files for their project (codesearch?), so it's not just the release scripts that rely on this behavior.

Patch-set: 5