Update patch set 1
Patch Set 1:

Because this isn't a git built-in, it won't autocomplete branch names, so it will be easy to typo branch names accidentally. If someone does, this will crash when it runs the actual archive command. It would be really good for this to check the existence of the provided branches, and print a warning + ignore missing ones.

Patch-set: 1
Reviewer: Gerrit User 1002100 <1002100@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>