Update patch set 2
Patch Set 2:

The reason for this is that if we treat a DEPS file as a .gclient file, and the DEPS file has recursedeps=["src/foo"] it means that we should only recurse into "src/foo" (as opposed to recursing into all deps, which is what .gclient files do).

I'm replacing the recursion_limit in GClient with recursedeps. By default it adds all solutions in the .gclient file, but with a DEPS file it would only add the solutions listed in recursedeps.

I got carried away and removed recursion_override. I hope we can get rid of it, since recursedeps is supposed to replace it anyway.

Patch-set: 2
Reviewer: Gerrit User 1167909 <1167909@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0