Update patch set 1
Patch Set 1:

(3 comments)

I don't really have concrete objections to this. It seems a bit odd to me in that I think it behaves more like a dep than not & the shared root is more of an implementation detail, but if y'all strongly think that it should be this way, I'm ok with it. It obviously makes the code more concise.

I am very concerned about this in light of the recent history of gclient breakages and their impact on chromium. How urgent is this? Does it need to land before we address some of the things we're thinking about doing to validate gclient before rolling it into chromium?

Patch-set: 1
Reviewer: Gerrit User 1181690 <1181690@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>