Update patch set 6
Patch Set 6:

.ninja_log contains all build-step times since the last gn clean. The times in .ninja_log are all relative to when ninja starts. So, if you do a build and it just links base.dll and that takes 1.1 s and if you do another build that just links base.dll and it takes 1.2 s then the existing logic will not realize that there were two separate builds, so the post-build summary will be wrong.

This doesn't happen all the time because the existing script notices when one build-step end time is less than the previous one and uses this as an indication that a new build was started.

This change makes the build-boundary detection much more reliable (not 100%, but probably 99.9%).

Patch-set: 6
Reviewer: Gerrit User 1128439 <1128439@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>