Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > We definitely don't want bot_update to be using gclient's "managed mode". We want to get rid of managed mode entirely, in fact. gclient should be in charge of making sure that deps repos match the sha1s pinned by their parent's DEPS files; it shouldn't be in the business of updating the top-level repos themselves unless explicitly asked to (e.g. by someone passing a --revision flag naming the top-level repo).
> 
> Yes, and this is how gclient works (see [1]). I think there is no need to have this in bot_update.
> 
> [1] https://chromium.googlesource.com/chromium/tools/depot_tools/+/master/gclient.py#1482

I'm not yet convinced that this is the right thing to do. In particular, the test change scares me, because it is replacing "foo@unmanaged" (aka "don't touch foo/") with "foo@DNE" (aka "try to sync foo/ to DNE"). Which is not the behavior change we want.

In particular, the comment at line 924 of bot_update.py suggests that the code there is working around a bug or weird corner case in gclient. I would like to be convinced that that weird corner case no longer exists before removing the code which works around it.

Patch-set: 1
Reviewer: Gerrit User 1002100 <1002100@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>