Use cwd from context instead of kwargs
It is not possible to use cwd parameter anymore, since none of the method called
by this module define cwd explicitly and low-level modules such as api.step do
not support it anymore, resulting in a TypeError exception due to unexpected
keyword argument when passing cwd into api.gclient.checkout(), e.g. see
https://ci.chromium.org/swarming/task/3e74a52905a71110. Given that any users of
this functionality are broken anyway, it should be safe to make this change.

R=tandrii@chromium.org

Bug: 793687
Change-Id: Iadf6bf91d8d41dbc6751a848bb69d7922a206ade